+++ docs/examples/datadog_exporter/server.py
@@ -21,7 +21,7 @@ from opentelemetry.exporter.datadog import (
 )
 from opentelemetry.exporter.datadog.propagator import DatadogFormat
 from opentelemetry.propagators.composite import CompositeHTTPPropagator
-from opentelemetry.propagators.util import (
+from opentelemetry.propagate import (
     get_global_textmap,
     set_global_textmap,
 )
