+++ docs/getting-started.rst
@@ -195,7 +195,7 @@ Following the installation of the package containing the b3 propagator, configur
 
 .. code-block:: python
 
-    from opentelemetry.propagators.util import set_global_textmap
+    from opentelemetry.propagate import set_global_textmap
     from opentelemetry.propagators.b3 import B3Format
 
     set_global_textmap(B3Format())
